Added in periodica updates for the stacks
DRYed up TwitterUsersController and TwitterUser and added in the
ability for a since_id to be passed into them. This doesn't seem
to be working at the moment, i think Ziggy may not be respecting the
arguements somewhere with it's caching.

Various other changes to the way helpers work. Probably ripe for refactoring